requirements.txt
pms5003 == 0.0.5
adafruit-io == 2.4.0
python-dotenv == 0.12.0
